+++ CHANGELOG.md
@@ -4,7 +4,8 @@ _Not yet on NuGet..._
 * Multiplot: Use new shared axis manager. `Multiplot.ShareY()` is now `Multiplot.SharedAxes.ShareY()` as seen in demo application. (#4738)
 * Multiplot: Moved state management and pixel lookup logic based on previous renders into the `IMultiplot.LastRender` property (#4738)
 * Multiplot: Added `Subplots` property for plot collection management and replaced `IMultiplot.Count` with `Multiplot.Count()` (#4738)
-* Axes: Added `IAxis.Collapse()` helper method to disable tick generation and set size to zero for a given axis
+* Axes: Added `IAxis.Collapse()` helper method to disable tick generation and set size to zero for a given axis (#4738)
+* Multiplot: Added `CollapseVertically()` to collapse space between vertically stacked plots (#4738)
 
 ## ScottPlot 5.0.53
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2025-01-12_